Restructure admin settings components by page
