+++ frontend_project_fund/app/admin/components/settings/funds_config/DeleteConfirmDialog.js
