+++ frontend_project_fund/app/admin/components/settings/index.js
@@ -1,8 +1,9 @@
-export { default } from "./main/FundSettingsContent";
+export { default } from "./FundSettingsContent";
 
-export { default as FundSettingsContent } from "./main/FundSettingsContent";
-export { default as AnnouncementManager } from "./announcements/AnnouncementManager";
-export { default as FundManagementTab } from "./tabs/FundManagementTab";
-export { default as YearManagementTab } from "./tabs/YearManagementTab";
-export { default as RewardConfigManager } from "./tabs/RewardConfigManager";
-export { default as SystemConfigSettings } from "./tabs/SystemConfigSettings";
+export { default as FundSettingsContent } from "./FundSettingsContent";
+export { default as StatusBadge } from "./StatusBadge";
+export { default as AnnouncementManager } from "./announcement_config/AnnouncementManager";
+export { default as FundManagementTab } from "./funds_config/FundManagementTab";
+export { default as YearManagementTab } from "./years_config/YearManagementTab";
+export { default as RewardConfigManager } from "./reward_config/RewardConfigManager";
+export { default as SystemConfigSettings } from "./system_config/SystemConfigSettings";
